backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2

# Database
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
pgvector==0.3.4

# Supabase
supabase==2.9.1

# AI / RAG
llama-index==0.11.14
llama-index-vector-stores-postgres==0.2.5
llama-index-embeddings-huggingface==0.3.1
llama-index-llms-groq==0.2.0
sentence-transformers==3.1.1
groq==0.11.0
openai==1.51.0
pyjwt==2.9.0

# Task Queue
celery[redis]==5.4.0
redis==5.1.1

# PDF Processing
pdfplumber==0.11.4
weasyprint==62.3

# Payments
paystack-python==2.0.0

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
orjson==3.10.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
